rename modules + cleanup
