+++ maitake/src/sync/wait_map/tests.rs
